requirements.txt
pillow~=8.4.0